Changes on TrafficEngineeringMetric

+++ src/es/tid/ospf/ospfv2/lsa/tlv/subtlv/TrafficEngineeringMetric.java
@@ -52,7 +52,7 @@ public class TrafficEngineeringMetric extends OSPFSubTLV {
 			}		
 	}
 
-	public float getLinkMetric() {
+	public long getLinkMetric() {
 		return linkMetric;
 	}
